flask/requirements.txt
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
uWSGI==2.0.20
Werkzeug==2.1.2
zipp==3.8.0
Flask-Cors==3.0.10
Flask-Injector==0.13.0
flask-restplus==0.13.0
frontend==0.0.3
future==0.18.2
h11==0.13.0
helpers==0.2.0
httplib2==0.20.4
idna==3.3
importlib-resources==5.7.1
injector==0.19.0
isodate==0.6.1
itsdangerous==2.1.2
jsonschema==4.4.0
lxml==4.8.0
MarkupSafe==2.1.1
networkx==2.8
nibabel==3.2.2
nipype==1.7.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pathlib==1.0.1
pdf2image==1.16.0
Pillow==9.2.0
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.19.6
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyxnat==1.4
rdflib==6.1.1
requests==2.27.1
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
starlette==0.20.0
traits==6.3.2
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
zipp==3.8.0
aiofiles==0.8.0
aniso8601==9.0.1
anyio==3.5.0
asgiref==3.4.1
attrs==21.4.0
cattrs==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
ci-info==0.2.0
configobj==5.0.6
configparser==5.2.0
etelemetry==0.3.0
exceptiongroup==1.0.0rc5
filelock==3.6.0

